requirements.txt
# === Core Django & Database ===
Django==5.1.6
dj-database-url==3.0.1
psycopg2-binary==2.9.11
sqlparse==0.5.3

# === Storage ===
supabase==2.8.0

# === Server & Deployment ===
gunicorn==23.0.0
python-decouple==3.8
python-dotenv==1.1.1

# === Frontend/Styling (Tailwind, optional) ===
django-tailwind==4.2.0
django-browser-reload==1.18.0

# === Utilities ===
pillow==11.3.0
requests==2.32.3
tzdata==2025.1
whitenoise==6.11.0

